feat: complete the app — content pack, all UI features, E2E, hardening
Seed content (content/core-en, CC0): 20 topics, 12 speaking topic sets
(48 Part 1/2/3 cards), 16 writing prompts incl. 6 Academic Task 1 chart
specs, 2 full Academic reading tests (6 passages, 80 keyed questions),
a 4-part listening test (40 questions), 343 vocabulary entries and 20
minimal pairs. All original; validates with checksums and seeds on a
fresh install.

UI: reading, listening, vocabulary, progress and onboarding built;
speaking and writing finished. All nine features register through the
route auto-discovery seam and render with zero console errors.

Fix a real data race in the writing autosave: save() returned early
whenever the ten-second tick was still in flight, so the pre-check GET
raced the PATCH and could report "you wrote 0 words" over a finished
essay. Callers now await the tick in progress and re-flush if the text
changed. Covered by three regression tests.

Verified: 558 pytest, 177 vitest, 14 Playwright specs driving the real
UI, tsc clean, production build clean. Reading scoring checked against
the official Academic table (40/40 -> 9.0, 20/40 -> 5.5, 0/40 -> 2.0).

+++ sidecar/bandready/scoring/rubrics.py
@@ -335,7 +335,7 @@ def descriptors_for(skill: str, criterion: str) -> dict[int, str]:
 def descriptor(skill: str, criterion: str, band: float) -> str:
     """One descriptor cell. Bands outside 4–9 clamp to the nearest paraphrased row."""
     rows = descriptors_for(skill, criterion)
-    whole = int(round(float(band)))
+    whole = round(float(band))
     whole = min(9, max(4, whole))
     return rows[whole]
